+++ collections/codec/int.go
@@ -7,9 +7,11 @@ import (
 	"strconv"
 )
 
-func NewInt64Key[T ~int64]() KeyCodec[T] { return int64Key[T]{} }
+func NewInt64Key[T ~int64]() NameableKeyCodec[T] { return int64Key[T]{} }
 
-type int64Key[T ~int64] struct{}
+type int64Key[T ~int64] struct {
+	name string
+}
 
 func (i int64Key[T]) Encode(buffer []byte, key T) (int, error) {
 	binary.BigEndian.PutUint64(buffer, (uint64)(key))
@@ -64,11 +66,22 @@ func (i int64Key[T]) SizeNonTerminal(_ T) int {
 	return 8
 }
 
-func NewInt32Key[T ~int32]() KeyCodec[T] {
+func (i int64Key[T]) WithName(name string) NamedKeyCodec[T] {
+	i.name = name
+	return i
+}
+
+func (i int64Key[T]) Name() string {
+	return i.name
+}
+
+func NewInt32Key[T ~int32]() NameableKeyCodec[T] {
 	return int32Key[T]{}
 }
 
-type int32Key[T ~int32] struct{}
+type int32Key[T ~int32] struct {
+	name string
+}
 
 func (i int32Key[T]) Encode(buffer []byte, key T) (int, error) {
 	binary.BigEndian.PutUint32(buffer, (uint32)(key))
@@ -120,4 +133,13 @@ func (i int32Key[T]) DecodeNonTerminal(buffer []byte) (int, T, error) {
 
 func (i int32Key[T]) SizeNonTerminal(_ T) int {
 	return 4
+}
+
+func (i int32Key[T]) WithName(name string) NamedKeyCodec[T] {
+	i.name = name
+	return i
+}
+
+func (i int32Key[T]) Name() string {
+	return i.name
 }